Camera2: Fix flash item status incorrect when turn on/off AE-Bracket
The popup should not be set null when the popup is showing, which
will lead to the view(popup) could not be update by application.
Example: when turn on/off AE-Bracket could not disable flash item
after popup is set null

Don't set popup null when the popup is showing.
[Reference similar code in VideoMenu]

CRs-Fixed: 674423

Change-Id: Id8c5772f3504014ee6591cc16dddcacbf5325629