Camera2: fix flash light turn on after resume camera activity
If user choose an app to open video, onResume will be called after
the choose dialog is dismiss. The function opencamera will be called
and the value mPreviewFocused will be set true, which lead to flash
light turn on when setCameraParameters.

The value mPreviewFocused won't be true if not in Camera Preview, we
can use the visible of mCameraControls to judge.

CRs-Fixed: 680450

Change-Id: Ia04ae0fb77f05fb991e5262f86cbe6f308db7709