While stopping FD, pause UI to avoid further drawing with stale data.
When user disables FD from settings, app will stop FD and sets the
callback listeners to null. By the time callbacks are set to null,
framework had posted the data and it was not received in the app yet.
So even though application clears the UI after setting null callbacks
one set of data is getting drawn and it will not be cleared at all.
Present change will pause the UI as soon as FD is disabled. When FD
is enabled again, photoUI resumes drawing by default so not calling
resume again.

CRs-Fixed: 712885
Change-Id: Iac3887eca6bc002439092390c47c25ae3b216b26