Restart face detection after caturing image in zsl mode.
When user does touch to focus face detection will restated after a delay.
But if user initiate a take picture in between all the pending messages
are cleared so face detection will not start automatically. This happens
only in ZSL has preview will not be restarted. This change restarts the
face detection if it is not started.

CRs-Fixed: 712889
Change-Id: I5cda7a3c2e7c8352cf6d9e423ca06eb63f6d6265