Camera2: correct image size recorded in media provider db
Photos taken in camcorder can't be previewed on PC via MTP mode. The
root cause is that, the actual image file size does not always match
the _size column in media provider database. During the saving process,
EXIF data are rewritten, but the size sent to media provider is not
updated accordingly.

Record bytes written in ExifOutputStream and OrderedDataOutputStream so
that after rewritting, the correct image size could be sent to media
provider.

Change-Id: I74d91e86d712c121eb50daecaf879ab4f5be97fc
CRs-Fixed: 711522